[ISSUE-358] Address review: inline labeled predicate logic into UDFs
The geaflow-dsl-common/function package is for general-purpose interfaces,
so the labeled predicate logic should live in the UDFs rather than a separate
helper class there.

- Inline the label-matching and three-valued-null logic (plus a private
  getLabel helper) directly into IsLabeled; IsNotLabeled delegates to an
  IsLabeled instance and negates, keeping label extraction in one place.
- Remove LabeledPredicateFunctions and its test from geaflow-dsl-common.
- Move the unit test to geaflow-dsl-plan as LabeledPredicateTest (same
  udf/table/other package as the UDFs, matching PropertyExistsTest), testing
  through the UDF eval methods.
