+++ geaflow/geaflow-dsl/geaflow-dsl-plan/src/main/java/org/apache/geaflow/dsl/udf/table/other/IsLabeled.java
@@ -22,22 +22,25 @@ package org.apache.geaflow.dsl.udf.table.other;
 import org.apache.geaflow.dsl.common.data.RowEdge;
 import org.apache.geaflow.dsl.common.data.RowVertex;
 import org.apache.geaflow.dsl.common.function.Description;
-import org.apache.geaflow.dsl.common.function.LabeledPredicateFunctions;
 import org.apache.geaflow.dsl.common.function.UDF;
 
 /**
  * UDF implementation for the ISO-GQL IS LABELED predicate.
  *
- * <p>Implements ISO-GQL Section 19.9: &lt;labeled predicate&gt;
+ * <p>Implements ISO-GQL Section 19.9: &lt;labeled predicate&gt;, which tests whether a
+ * graph element (vertex or edge) has a given label.
  *
  * <p><b>Syntax:</b></p>
  * <pre>
  *   IS_LABELED(element, label)
  * </pre>
  *
- * <p><b>Semantics:</b></p>
- * Returns TRUE if the vertex or edge has the given label, FALSE if not, or NULL if either
- * operand is NULL.
+ * <p><b>Semantics (ISO-GQL three-valued logic):</b></p>
+ * <ul>
+ *   <li>If the element or the label is null, the result is Unknown (null).</li>
+ *   <li>If the element's label equals the given label, the result is True.</li>
+ *   <li>Otherwise, the result is False.</li>
+ * </ul>
  *
  * <p><b>Example:</b></p>
  * <pre>
@@ -62,20 +65,37 @@ public class IsLabeled extends UDF {
      *         operand is null
      */
     public Boolean eval(Object elementValue, Object labelValue) {
-        return LabeledPredicateFunctions.isLabeled(elementValue, labelValue);
+        // ISO-GQL Rule: If element or label is null, result is Unknown (null).
+        if (elementValue == null || labelValue == null) {
+            return null;
+        }
+        String elementLabel = getLabel(elementValue);
+        return elementLabel != null && elementLabel.equals(labelValue.toString());
     }
 
     /**
      * Type-specific overload for vertices.
      */
     public Boolean eval(RowVertex vertex, String label) {
-        return LabeledPredicateFunctions.isLabeled(vertex, label);
+        return eval((Object) vertex, label);
     }
 
     /**
      * Type-specific overload for edges.
      */
     public Boolean eval(RowEdge edge, String label) {
-        return LabeledPredicateFunctions.isLabeled(edge, label);
+        return eval((Object) edge, label);
+    }
+
+    private static String getLabel(Object elementValue) {
+        if (elementValue instanceof RowVertex) {
+            return ((RowVertex) elementValue).getLabel();
+        }
+        if (elementValue instanceof RowEdge) {
+            return ((RowEdge) elementValue).getLabel();
+        }
+        throw new IllegalArgumentException(
+            "First operand of labeled predicate must be a vertex or an edge, got: "
+                + elementValue.getClass().getName());
     }
 }
